app/requirements.txt
mecab-python3==1.0.1
wordcloud==1.8.1